Refactor das camadas Core e Aluno
