+++ src/backend/alunos-api/src/Alunos.Application/Commands/AtualizarPagamento/AtualizarPagamentoMatriculaCommand.cs
@@ -1,6 +1,6 @@
 ﻿using Core.Messages;
 
-namespace Core.Mediator.AlunoCommands;
+namespace Alunos.Application.Commands.AtualizarPagamento;
 public class AtualizarPagamentoMatriculaCommand : CommandRaiz
 {
     public Guid AlunoId { get; init; }
