+++ src/backend/alunos-api/src/Alunos.Application/Commands/ConcluirCurso/ConcluirCursoCommand.cs
@@ -1,7 +1,7 @@
 ﻿using Core.Messages;
 using Core.SharedDtos.Conteudo;
 
-namespace Core.Mediator.AlunoCommands;
+namespace Alunos.Application.Commands.ConcluirCurso;
 public class ConcluirCursoCommand : CommandRaiz
 {
     public Guid AlunoId { get; init; }
