+++ src/backend/alunos-api/src/Alunos.Application/Commands/ConcluirCurso/ConcluirCursoCommandHandler.cs
@@ -2,7 +2,6 @@
 using Alunos.Domain.Interfaces;
 using Core.Communication;
 using Core.Mediator;
-using Core.Mediator.AlunoCommands;
 using Core.Messages;
 using Core.SharedDtos.Conteudo;
 using MediatR;
