+++ src/backend/alunos-api/src/Alunos.Application/Commands/ConcluirCurso/ConcluirCursoCommandValidator.cs
@@ -1,5 +1,4 @@
-﻿using Core.Mediator.AlunoCommands;
-using FluentValidation;
+﻿using FluentValidation;
 
 namespace Alunos.Application.Commands.ConcluirCurso;
 public class ConcluirCursoCommandValidator : AbstractValidator<ConcluirCursoCommand>
