+++ src/backend/alunos-api/src/Alunos.Application/Commands/MatricularAluno/MatricularAlunoCommandValidator.cs
@@ -1,5 +1,4 @@
-﻿using Core.Mediator.AlunoCommands;
-using FluentValidation;
+﻿using FluentValidation;
 
 namespace Alunos.Application.Commands.MatricularAluno;
 public class MatricularAlunoCommandValidator : AbstractValidator<MatricularAlunoCommand>
