+++ src/backend/alunos-api/src/Alunos.Application/Commands/RegistrarHistoricoAprendizado/RegistrarHistoricoAprendizadoCommandValidator.cs
@@ -1,5 +1,4 @@
-﻿using Core.Mediator.AlunoCommands;
-using FluentValidation;
+﻿using FluentValidation;
 
 namespace Alunos.Application.Commands.RegistrarHistoricoAprendizado;
 public class RegistrarHistoricoAprendizadoCommandValidator : AbstractValidator<RegistrarHistoricoAprendizadoCommand>
