+++ src/backend/alunos-api/src/Alunos.Application/Commands/SolicitarCertificado/SolicitarCertificadoCommand.cs
@@ -1,6 +1,6 @@
 ﻿using Core.Messages;
 
-namespace Core.Mediator.AlunoCommands;
+namespace Alunos.Application.Commands.SolicitarCertificado;
 public class SolicitarCertificadoCommand : CommandRaiz
 {
     public Guid AlunoId { get; private set; }
